+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix MacBookPro5,4 audio support
 - Increment binaries
   - AppleALC f2889fc (1.6.1 rolling - 05-03-2021)
+- Support custom CPU names in About This Mac
 
 ## 0.1.2
 - Fix IDE support on 2008 era MacBooks, iMacs and Xserves